feat(后端):实现教练查询待审核学员，教练审核学员，学员预约课程功能

+++ backend/src/main/kotlin/io/github/shaoyuanyu/ttts/persistence/course/CourseTable.kt
@@ -40,7 +40,7 @@ object CourseTable : UUIDTable("courses") {
         toDb = { it.name }
     )
 
-    val price: Column<Int> = integer("price")
+    val price: Column<Float> = float("price")
 
     // PENDING, PAID, REFUNDED
     val paymentStatus: Column<String> = varchar("payment_status", 32).default("PENDING")
